+++ NEWS.md
@@ -1,4 +1,9 @@
 # mlfinance 0.0.1
 
-* Added a `NEWS.md` file to track changes to the package.
-* New cusum_filter function for filtering barsbased on cusum filter
+* New `daily_volatility` function for calculating daily variance from intraday data
+* New `cusum_filter` function for filtering
+* New `get_events` function for generating triple barrier events
+* New `get_labels` function for generating triple barrier labels
+* New `drop_labels` function for droping labels
+
+---